+++ src/components/templates/admin/member-management/member-list/index.tsx
@@ -1,171 +1,115 @@
-import { useDisclosure } from "@chakra-ui/react";
-import ModalForm from "@src/components/organisms/modal";
-import { useEffect, useState } from "react";
-import { useFormHook } from "@src/hooks/useFormhook";
-import * as yup from "yup";
-import { createPhoneNumberSchema } from "@src/utility/phoneValidation";
-import { IParams } from "@src/interface/params";
-import { Pagination } from "@src/components/organisms/table";
-import { IMember } from "@src/interface/member-management";
-import MemberTable from "../member-table";
-import { CreateMemberManagementForm } from "../member-add";
-import { usePropertyList } from "@src/constant/usePropertyList";
-import { useMutation, useQueryClient } from "react-query";
-import { toastFail, toastSuccess } from "@src/service/service-toast";
-import { createMember } from "@src/service/member-management";
-import { AxiosError } from "axios";
+import { Stack } from "@chakra-ui/react";
+import Switch from "@src/components/atoms/Switch";
+import BasicTable from "@src/components/molecules/table";
+import TableHeadings from "@src/components/molecules/table-heading";
+import TableActions from "@src/components/molecules/table/TableActions";
+import { usePageParams } from "@src/components/organisms/layout";
+import { NAVIGATION_ROUTES } from "@src/routes/routes.constant";
+import { useMemo } from "react";
+import { useNavigate } from "react-router-dom";
+import { CellProps } from "react-table";
 
 interface IMemberProps {
-  tableData: IMember[];
-  tableDataFetching: boolean;
-  _pageSizeChange: (limit: number) => void;
-  _pageChange: (page: number) => void;
-  paginatedData: IMember[];
-  pageParams: IParams;
+  setUpdateId: any;
+  setIsUpdate: any;
+  onMemberModalOpen: any;
+  onCloseHandler: any;
+  data: any;
+  isLoading: any;
 }
 
-const defaultValues = {
-  fullName: "",
-  email: "",
-  phoneNumber: "",
-  nationalityId: "",
-  dateOfBirth: "",
-  membershipTierId: "",
-  isActive: false,
-};
-const validationSchema = yup.object().shape({
-  fullName: yup.string().required("Full Name is required"),
-  email: yup.string().required("Email is required"),
-  phoneNumber: createPhoneNumberSchema(),
-  nationalityId: yup.string().required("Nationality is required"),
-  propertyId: yup.string().required("Property Name is required"),
-  dateOfBirth: yup.string().required("DOB is required"),
-});
-
 const MemberManagementList: React.FC<IMemberProps> = ({
-  tableData,
-  tableDataFetching,
-  _pageSizeChange,
-  _pageChange,
-  paginatedData,
-  pageParams,
+  setUpdateId,
+  setIsUpdate,
+  onMemberModalOpen,
+  onCloseHandler,
+  data,
+  isLoading,
 }) => {
-  const [isUpdate, setIsUpdate] = useState(false);
-  const [updateId, setUpdateId] = useState("");
+  const { pageParams } = usePageParams();
+  const navigate = useNavigate();
 
-  const propertyList = usePropertyList();
-  const {
-    isOpen: isMemberOpen,
-    onOpen: onMemberModalOpen,
-    onClose: onMemberModalClose,
-  } = useDisclosure();
+  const columns = useMemo(
+    () => [
+      {
+        header: "S.N",
+        accessorFn: (_: any, index: number) =>
+          (pageParams.page - 1) * pageParams.limit + (index + 1),
+      },
 
-  const { handleSubmit, register, errors, reset, setValue } = useFormHook({
-    validationSchema,
-    defaultValues,
-  });
+      {
+        header: "Full Name",
+        accessorKey: "fullName",
+      },
+      {
+        header: "Email",
+        accessorKey: "email",
+      },
+      {
+        header: "Phone Number",
+        accessorKey: "phoneNumber",
+      },
+      {
+        header: "Nationality",
+        accessorKey: "nationality",
+      },
+      {
+        header: "Tier",
+        accessorKey: "membershipTierName",
+      },
+      {
+        header: "Is Blocked?",
+        accessorKey: "isBlocked",
+        cell: ({ row }: CellProps<{ isBlocked: boolean }>) => {
+          const status = row?.original?.isBlocked;
+          return status === true ? (
+            <Switch value={true} variant="red" />
+          ) : (
+            <Switch disabled />
+          );
+        },
+      },
+      {
+        header: "Action",
+        cell: ({ row }: CellProps<{ id: string }>) => {
+          const onEdit = () => {
+            setUpdateId(row.original.id);
+            setIsUpdate(true);
+            onMemberModalOpen();
+          };
+          const onView = () => {
+            navigate(NAVIGATION_ROUTES.PROFILE_DETAIL, {
+              state: row.original,
+            });
+          };
 
-  useEffect(() => {
-    if (isUpdate && updateId) {
-      const data = tableData.find((x: IMember) => x.id === updateId);
-      reset({
-        ...data,
-      });
-    }
-  }, [isUpdate, updateId, tableData]);
-  const queryClient = useQueryClient();
-  const { mutate, isLoading } = useMutation(createMember, {
-    onSuccess: (response) => {
-      toastSuccess(response?.data?.message || "Congratulations!");
-      queryClient.refetchQueries("member_management");
-      onCloseHandler();
-    },
-    onError: (error: AxiosError<{ message: string }>) => {
-      toastFail(error?.response?.data?.message || "Something went wrong");
-    },
-  });
-
-  const onCloseHandler = () => {
-    reset(defaultValues);
-    setUpdateId("");
-    setIsUpdate(false);
-    onMemberModalClose();
-  };
-
-  const onSubmitHandler = async (data: IMember) => {
-    if (updateId) {
-      mutate({
-        id: updateId,
-        fullName: data.fullName,
-        email: data.email,
-        phoneNumber: data.phoneNumber,
-        dateOfBirth: data.dateOfBirth,
-        nationalityId: data.nationalityId,
-        propertyId: data?.propertyId,
-        isBlocked: data.isBlocked,
-        membershipTierId: data.membershipTierId,
-        roleId: "2",
-      });
-    } else {
-      mutate({
-        id: "0",
-        fullName: data.fullName,
-        email: data.email,
-        dateOfBirth: data.dateOfBirth,
-        phoneNumber: data.phoneNumber,
-        nationalityId: data.nationalityId,
-        propertyId: data.propertyId,
-        roleId: "2",
-      });
-    }
-  };
+          return (
+            <Stack alignItems={"flex-start"}>
+              <TableActions onEdit={onEdit} onView={onView} />
+            </Stack>
+          );
+        },
+      },
+    ],
+    [pageParams]
+  );
 
   return (
     <>
-      <MemberTable
-        tableDataFetching={tableDataFetching}
-        paginatedData={paginatedData}
-        pageParams={pageParams}
-        title="Filter By"
+      <TableHeadings
         btnText="Add Member"
         CurrentText="Member List"
         onAction={() => {
           onCloseHandler();
           onMemberModalOpen();
         }}
-        onEditData={(id: string) => {
-          setUpdateId(id);
-          setIsUpdate(true);
-          onMemberModalOpen();
-        }}
       />
-
-      <Pagination
-        enabled={true}
-        queryPageIndex={pageParams.page}
-        queryPageSize={pageParams.limit}
-        totalCount={tableData?.length || 0}
-        pageChange={_pageChange}
-        pageSizeChange={_pageSizeChange}
-      />
-      <ModalForm
-        isModalOpen={isMemberOpen}
-        title={isUpdate ? "Update Member" : "Add Member"}
-        onCloseModal={onMemberModalClose}
-        resetButtonText={"Cancel"}
+      <BasicTable
+        data={data?.data || []}
+        columns={columns}
         isLoading={isLoading}
-        submitButtonText={isUpdate ? "Update Member" : "Add Member"}
-        submitHandler={handleSubmit(onSubmitHandler)}
-        showFooter={true}
-      >
-        <CreateMemberManagementForm
-          register={register}
-          errors={errors}
-          propertyList={propertyList}
-          setValue={setValue}
-          id={updateId}
-        />
-      </ModalForm>
+        totalPages={data?.totalPages}
+      />
     </>
   );
 };